Rename folder and imports /(components) to  /components
